gathering: Rename gather package to gathering
We want to add a gather command depending on the validation package, but
the validation package depends on the gather package. Move the gather
implementation to new gathering package. This validation and testing
packages use it.
