+++ pkg/gathering/gathering.go
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: The RamenDR authors
 // SPDX-License-Identifier: Apache-2.0
 
-package gather
+package gathering
 
 import (
 	"path/filepath"
