+++ pkg/validation/mock.go
@@ -5,7 +5,7 @@ package validation
 
 import (
 	"github.com/ramendr/ramen/e2e/types"
-	"github.com/ramendr/ramenctl/pkg/gather"
+	"github.com/ramendr/ramenctl/pkg/gathering"
 )
 
 type ContextFunc func(Context) error
@@ -15,7 +15,7 @@ type ContextFunc func(Context) error
 type Mock struct {
 	ValidateFunc              ContextFunc
 	ApplicationNamespacesFunc func(ctx Context, drpcName, drpcNamespace string) ([]string, error)
-	GatherFunc                func(ctx Context, clsuters []*types.Cluster, namespaces []string, outputDir string) <-chan gather.Result
+	GatherFunc                func(ctx Context, clsuters []*types.Cluster, namespaces []string, outputDir string) <-chan gathering.Result
 }
 
 var _ Validation = &Mock{}
@@ -42,14 +42,14 @@ func (m *Mock) Gather(
 	clusters []*types.Cluster,
 	namespaces []string,
 	outputDir string,
-) <-chan gather.Result {
+) <-chan gathering.Result {
 	if m.GatherFunc != nil {
 		return m.GatherFunc(ctx, clusters, namespaces, outputDir)
 	}
 
-	results := make(chan gather.Result, len(clusters))
+	results := make(chan gathering.Result, len(clusters))
 	for _, cluster := range clusters {
-		results <- gather.Result{Name: cluster.Name, Err: nil}
+		results <- gathering.Result{Name: cluster.Name, Err: nil}
 	}
 	close(results)
 	return results
